==> 00c_feat_rnaelements.Rmd <==
---
title: "asthmaBiomarkers_discoveryDataProcessing_nanoString"
subtitle: "input: nanostring elements files & demographics; output: well formatted data and meta col files"
output:
  html_document:
    theme: paper
    highlight: zenburn
    number_sections: true
    toc: true
    toc_float: 
      collapsed: true
      smooth_scroll: true
---

```{r, echo = FALSE, results='hide', message=FALSE, warning=FALSE}
knitr::opts_chunk$set(warning = FALSE, message = FALSE)
```


```{r}
## logistics
root = "~/projects/asthma" # root directory, used for _dirs.R
source(paste0(root, "/src/_dirs.R"))
source(paste0(root, "/src/_func.R"))
source(paste0(root, "/src/_func-asthma.R"))
libr(append(pkgs(), c("pracma","NormqPCR","mixOmics","amritr")))
```


```{r}
load(rnaseq_to_ns_dir) # load list of biomarker candidates

# 1) HBA attenutation (Discovery cohort)

setName <- paste("Set", 1:length(hba_attenuation_discovery_dirs))
date <- c("2016-01-14","2016-01-16","2016-01-16", "2016-01-16") #date?
cartridgeLotNum <- rep("31020635468_exp02.2016", length(hba_attenuation_discovery_dirs))
Tag156LotNum <- rep("TS2877-1", length(hba_attenuation_discovery_dirs))
ext24TagLotNum <- rep("TS4006", length(hba_attenuation_discovery_dirs))
plateLotNum <- rep("41000274-0056_exp02.2016", length(hba_attenuation_discovery_dirs))
DatList <- list()
for(h in 1 : length(hba_attenuation_discovery_dirs)){
  files <- list.files(hba_attenuation_discovery_dirs[h], full.names = TRUE)
  dat <- lapply(as.list(files), function(i){
    rccToDat(fileName = i)
  }) %>% do.call(rbind, .)
  dat$Set <- setName[h]
  dat$Date <- date[h]
  dat$cartridgeLotNum <- cartridgeLotNum[h]
  dat$Tag156LotNum <- Tag156LotNum[h]
  dat$ext24TagLotNum <- ext24TagLotNum[h]
  dat$plateLotNum <- plateLotNum[h]
  DatList[[h]] <- dat
}
Dat <- do.call(rbind, DatList)
write.csv(Dat, hba_attenuation_discovery_dir)

#------------
# import data ### NO METADATA!!
#------------
sampleMap0 <- read.csv(hba_attenuation_discovery_demo_dir)
sampleMap <- c(paste0("E", sampleMap0$EorV_number))
names(sampleMap) <- paste(sampleMap0$fileName, unlist(lapply(strsplit(as.character(sampleMap0$Set_lane), "_"), function(i) i[2])), sep = "_")
data <- read.csv(hba_attenuation_discovery_data_dir, row.names = 1)
data$Sample <- sampleMap[gsub(".RCC", "", unlist(lapply(strsplit(as.character(data$fileName), "/"), function(i) i[length(i)])))]
GeneNames <- as.character(data$Name)
GeneNames[data$Accession == "NM_001788.5"] <- "SEPT7"
data$Name <- GeneNames
srcClass <- as.character(data$CodeClass)
srcClass[data$Name == "comp56964_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
srcClass[data$Name == "comp56975_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
srcClass[data$Name == "comp56957_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
data$CodeClass <- factor(srcClass)
asthmaDis <- data

#----------------------------------------------
# 2) HBA attenutation (confirmatory phase) #### NO DATA!
#----------------------------------------------
setName <- paste("Set", 1:length(hba_attenuation_confirmatory_data_dirs))
date <- rep(c("2016-02-10","2016-02-11"), each = 3)
cartridgeLotNum <- rep("31020635467_exp12.2016", length(hba_attenuation_confirmatory_data_dirs))
Tag156LotNum <- rep("TS2877-1", length(hba_attenuation_confirmatory_data_dirs))
ext24TagLotNum <- rep("TS4006", length(hba_attenuation_confirmatory_data_dirs))
plateLotNum <- rep("41000274-0056_exp02.2016", length(hba_attenuation_confirmatory_data_dirs))
DatList <- list()
for(h in 1:length(hba_attenuation_confirmatory_data_dirs)){
  files <- list.files(hba_attenuation_confirmatory_data_dirs[h], full.names = TRUE)
  dat <- lapply(as.list(files), function(i) rccToDat(i) ) %>% do.call(rbind, .)
  dat$Set <- setName[h]
  dat$Date <- date[h]
  dat$cartridgeLotNum <- cartridgeLotNum[h]
  dat$Tag156LotNum <- Tag156LotNum[h]
  dat$ext24TagLotNum <- ext24TagLotNum[h]
  dat$plateLotNum <- plateLotNum[h]
  DatList[[h]] <- dat
}
Dat <- do.call(rbind, DatList)
write.csv(Dat, hba_attenuation_confirmatory_data_dir)

#------------
# import data ### NO MAPPING FILE
#------------
mappingFile <- read.csv(hba_attenuation_confirmatory_demo_dir, row.names = 7)
data <- read.csv(hba_attenuation_confirmatory_data_dir, row.names = 1)
sampleMap <- gsub(".RCC", "", paste(unlist(lapply(strsplit(as.character(data$fileName), " "), function(i) i[3])),
                                    unlist(lapply(strsplit(as.character(data$fileName), "_"), function(i) i[8])), sep = "_"))
data$Sample <- mappingFile[sampleMap, "V.ID"]
GeneNames <- as.character(data$Name)
GeneNames[data$Accession == "NM_001788.5"] <- "SEPT7"
data$Name <- GeneNames
srcClass <- as.character(data$CodeClass)
srcClass[data$Name == "comp56964_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
srcClass[data$Name == "comp56975_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
srcClass[data$Name == "comp56957_c0_seq1_hk"] <- "Housekeeping_Code_Summary"
data$CodeClass <- factor(srcClass)
asthmaVal <- data

asthmaDat <- rbind(asthmaDis, asthmaVal)
asthmaDat$Cohort <- rep(c("January_2016", "February_2016"), c(nrow(asthmaDis), nrow(asthmaVal)))

## filter for biomarker candidates and house-keepers
ucscGenes <- as.character(ucscGenes_candidates$GenSym)
ucscGeneIso <- as.character(ucscGeneIso_candidates$GenSym)
ensembl <- as.character(ensembl_candidates$GenSym)
trinity <- as.character(trinity_candidates$GenSym)
all <- unique(c(ucscGenes, ucscGeneIso, ensembl, trinity))
asthmaDat <- subset(asthmaDat, Name %in% c(unique(grep("POS|NEG", asthmaDat$Name, value = TRUE)), all, houseKeepers))

```

# import demographics

```{r}
demo0 <- readRDS(meta_file_rnapc_dir)
demo0 <- demo0[as.character(demo0$EorV_number) %in% unique(asthmaDat$Sample), ]
rownames(demo0) <- as.character(demo0$EorV_number)
demo0$rnaseq_coreSet_biomarkerAnalysis <- as.character(demo0$rnaseq_coreSet_biomarkerAnalysis)

# fix site
demo0$SITE <- as.character(demo0$SITE)
demo0$SITE[demo0$calculated_Response == "Control"] <- "UBC"

controls <- subset(demo0, calculated_Response == "Control")

## Time period when these samples were collected
range(as.numeric(sapply(strsplit(as.character(demo0$AIC_YMD), "-"), function(i) i[1])))
```

## create discovery validation and repeatChallenges dataset

```{r}
demo1_dis <- subset(demo0, rnaseq_coreSet_biomarkerAnalysis == "Y")
demo1_dis$Allergen_cleanLabel <- as.character(demo1_dis$Allergen_cleanLabel)
demo1_val <- subset(demo0, rnaseq_coreSet_biomarkerAnalysis != "Y")
## 
## further remove controls
demo1_val <- subset(demo1_val, calculated_Response != "Control")
## remove replicates from discovery cohort
demo1_val <- demo1_val[!(demo1_val$NAME %in% demo1_dis$NAME), ]
## remove subjects that were challenged with allergens results/enrichr than that of the discovery cohort
demo1_val <- demo1_val[as.character(demo1_val$Allergen_cleanLabel) %in%  unique(as.character(demo1_dis$Allergen_cleanLabel)), ]
## remove subjects with repeat challenges
rmSubj <- names(table(as.character(demo1_val$NAME))[table(as.character(demo1_val$NAME)) > 1])
## remove replicates for final validation dataset
demo1_val <- demo1_val[!(demo1_val$NAME %in% rmSubj), ]

## different response
flippers <- names(rowSums(table(as.character(demo0$NAME), demo0$calculated_Response) != 0)[rowSums(table(as.character(demo0$NAME), demo0$calculated_Response) != 0) == 2])
demo1_val_repeat <- demo0[demo0$NAME %in% flippers, ]
table(as.character(demo1_val_repeat$NAME), demo1_val_repeat$calculated_Response)
# remove subjects challenges with different allergens
demo1_val_repeatDifferentResponse <- demo1_val_repeat[!(as.character(demo1_val_repeat$NAME) %in% names(rowSums(table(as.character(demo1_val_repeat$NAME), as.character(demo1_val_repeat$Allergen_cleanLabel)) != 0)[rowSums(table(as.character(demo1_val_repeat$NAME), as.character(demo1_val_repeat$Allergen_cleanLabel)) != 0) == 2])), ]

## same response
demo_sameResponse1 <- subset(demo0, calculated_Response != "Control")
dropFlippers <- rownames(table(as.character(demo_sameResponse1$NAME), demo_sameResponse1$calculated_Response))[rowSums(table(as.character(demo_sameResponse1$NAME), demo_sameResponse1$calculated_Response) == 0) == 1]
demo_sameResponse2 <- demo_sameResponse1[demo_sameResponse1$NAME %in% dropFlippers, ]
demo1_val_repeatSameResponse <- demo_sameResponse2[demo_sameResponse2$NAME %in% rownames(table(as.character(demo_sameResponse2$NAME), demo_sameResponse2$calculated_Response))[rowSums(table(as.character(demo_sameResponse2$NAME), demo_sameResponse2$calculated_Response) > 1) == 1], ]
## remove subjects challenged with different allergens
name.allergen <- table(as.character(demo1_val_repeatSameResponse$NAME), as.character(demo1_val_repeatSameResponse$Allergen_cleanLabel))
demo1_val_repeatSameResponse <- demo1_val_repeatSameResponse[demo1_val_repeatSameResponse$NAME %in% rownames(name.allergen)[rowSums(name.allergen == 0) == 4], ]
## keeps DRs only
demo1_val_repeatDR <- subset(demo1_val_repeatSameResponse, calculated_Response == "DR")

table(demo1_dis$calculated_Response)
table(demo1_val$calculated_Response)
table(demo1_val_repeatDifferentResponse$calculated_Response, as.character(demo1_val_repeatDifferentResponse$Allergen_cleanLabel))
table(demo1_val_repeatDR$calculated_Response, as.character(demo1_val_repeatDR$Allergen_cleanLabel),
  demo1_val_repeatDR$concealed_ID)

```

## filter nanoString to only contain samples from subjects I actually use in the analyses

```{r}
asthmaDat <- asthmaDat[asthmaDat$Sample %in% c(rownames(demo1_dis), rownames(demo1_val), rownames(demo1_val_repeatDifferentResponse), rownames(demo1_val_repeatDR)), ]

```

# Quality Control
## Binding density and FOVratio

```{r}
##################################################
## quality control
##################################################
qc <- asthmaDat %>% mutate(FOVratio = FovCounted_Lane_Attributes/FovCount_Lane_Attributes) %>% 
  dplyr::select(FOVratio, BindingDensity = BindingDensity_Lane_Attributes, Sample, Set, ID_Lane_Attributes, Cohort) %>%
  gather(QCVar, QCmetrics, -c(Sample, Set, ID_Lane_Attributes, Cohort)) %>% dplyr::group_by(Sample, QCVar, Set, Cohort) %>% 
  #distinct(Sample, Set, ID_Lane_Attributes, Cohort) %>% 
  dplyr::mutate(Sample_Lane = paste(Sample, ID_Lane_Attributes, sep="_"))
qc$Sample_Lane <- factor(qc$Sample_Lane, levels = unique(qc$Sample_Lane))

hline.data <- data.frame(yint = c(0.75, 0.05, 2.25),
  QCVar = c("FOVratio", rep("BindingDensity", 2)))
ggplot(data = qc, aes(x = Sample_Lane, y = QCmetrics)) + geom_point() + 
  facet_grid(QCVar~Cohort, scale = "free") + 
  geom_hline(data = hline.data, aes(yintercept = yint), lty = 2, col = "red") + 
  customTheme(sizeStripFont=15, xAngle=90, hjust = 1, 
    vjust = 0.5, xSize=5, ySize=15, xAxisSize=15, yAxisSize=15)

```

## Linearity of positive controls

```{r}
posConc <- c(128, 32, 8, 2, 0.5, 0.125)
names(posConc) <- paste0(paste("POS", LETTERS[1:6], sep = "_"), paste0("(", posConc, ")"))
posDat <- asthmaDat %>% filter(CodeClass == "Positive_Code_Summary") %>% 
  mutate(posConc = posConc[as.character(Name)],
    Sample_Lane = paste(Sample, ID_Lane_Attributes, sep="_"))

mods <- posDat %>% 
  group_by(ID_Lane_Attributes, Set, Cohort) %>% do(., r2 = summary(lm(Count ~ posConc, data = .))$r.squared)
r.squared <- data.frame(ID_Lane_Attributes = mods$ID_Lane_Attributes, R2 = round(unlist(mods$r2), 2),
  Set = mods$Set)
r.squared$x <- 20
r.squared$y <- 100
all(unlist(mods$r2) > 0.9)


```

## lower limit of detection

```{r}
#### Negative Controls
negSD <- asthmaDat %>% dplyr::filter(CodeClass == "Negative_Code_Summary") %>% group_by(fileName, Set, Sample, Cohort) %>% dplyr::summarise(negThres = mean(Count)+2*sd(Count))
pos0.5 = posDat %>% filter(posConc == 0.5)
all(levels(negSD$Sample) == levels(pos0.5$Sample))   # TRUE
negPos <- as.data.frame(rbind(negSD[, c("Sample", "Set", "negThres", "Cohort")], setNames(pos0.5[, c("Sample", "Set", "Count", "Cohort")], c("Sample", "Set", "negThres", "Cohort"))))
negPos$Type = rep(c("Neg(Mean+2SD)","Pos0.5fM"), each=nrow(negSD))
negPos$Sample <- factor(negPos$Sample, levels = unique(negPos$Sample)[order(as.numeric(gsub("E","",unique(negPos$Sample))))])
ggplot(negPos, aes(x = Sample, y = negThres, color = Type)) + geom_point() + ylab("Count") + 
  customTheme(sizeStripFont=15, xAngle=90, hjust = 1, vjust = 0.5, xSize=5, ySize=8, xAxisSize=5, yAxisSize=8) + 
  facet_grid(Cohort~.)

```

## Counts for each type of tag

```{r}
## 4) Counts for each type of tag
asthmaDat$CodeClass <- gsub("_Code_Summary", "", asthmaDat$CodeClass)
meanFeature <- asthmaDat %>% group_by(CodeClass, Name) %>% dplyr::summarise(mean = mean(Count)) %>% dplyr::arrange(mean)
asthmaDat$Name <- factor(as.character(asthmaDat$Name), levels = as.character(meanFeature$Name))

asthmaDat %>% group_by(fileName) %>% dplyr::summarise(TotalSum = sum(Count), Cohort=unique(Cohort), Sample=unique(Sample)) %>% 
  dplyr::select(TotalSum, Cohort) %>% 
  ggplot(aes(x = TotalSum, fill = Cohort)) + geom_density(alpha = 0.5) + 
  customTheme(sizeStripFont=15, xAngle=0, hjust = 0.5, vjust = 0.5, xSize=10, 
    ySize=10, xAxisSize=10, yAxisSize=10) +
  theme(legend.position = c(0.7,0.9))
## plot assay counts
Dat2 <- asthmaDat
Dat2$Name <- as.character(Dat2$Name)
Dat2$Name[Dat2$Name != "HBA2"] <- "results/enrichr genes"
Dat3 <- Dat2 %>% group_by(fileName, Name) %>% dplyr::summarise(TotalSum = sum(Count), Cohort=unique(Cohort)) %>% 
  dplyr::select(TotalSum, Cohort, Name) 

Dat3 %>% 
  ggplot(aes(x = Name, y = TotalSum)) + geom_boxplot() +
  facet_grid(.~Cohort) +
  customTheme(sizeStripFont=9, xAngle=0, hjust = 0.5, vjust = 0.5, xSize=7, 
    ySize=10, xAxisSize=10, yAxisSize=10) + 
  scale_y_log10(breaks = c(1e5,2e5,seq(0,2e6, 5e5))) + xlab("")+
  ylab("Total counts for each assay") +
  theme(legend.position = c(0.5,0.2), legend.title=element_blank(),
    legend.text=element_text(size=7), legend.key=element_blank())
```

## counts per each gene

```{r}
## plot assay counts per gene
ggplot(asthmaDat, aes(x = Name, y = Count, color = CodeClass)) + geom_point() + scale_y_log10() + 
  customTheme(sizeStripFont=15, xAngle=90, hjust = 1, vjust = 0.5, xSize=6, ySize=8, xAxisSize=8, yAxisSize=8) + 
  theme(legend.position = c(0.1, 0.85)) + xlab("CodeClass") +
  annotate("text", label = "HBA2", x = 100, y = 1500000, size = 4, colour = "black")

```

# Normalization
## positive control normalization

```{r}
table(as.character(asthmaDat$Sample), as.character(asthmaDat$Date))
posConc <- c(128, 32, 8, 2, 0.5, 0.125)
names(posConc) <- paste0(paste("POS", LETTERS[1:6], sep = "_"), paste0("(", posConc, ")"))
posDat <- asthmaDat %>% filter(CodeClass == "Positive") %>% 
    mutate(posConc = posConc[as.character(Name)],
      Sample_Lane = paste(Sample, ID_Lane_Attributes, sep="_"))
  
## Positive Control Normalization
posNF <- posDat %>% group_by(fileName) %>% 
  dplyr::summarise(posGeoMean = geometric.mean(Count)) %>% 
    dplyr::mutate(posNF = mean(posGeoMean)/posGeoMean)
all(posNF$posNF<3 & posNF$posNF > 0.3)
  
dataNorm <- full_join(asthmaDat, posNF, by = "fileName") %>% 
  mutate(posNorm = Count*posNF)
  
# check if the sum of the positive controls is the same for each assay
dataNorm %>% group_by(fileName) %>% filter(CodeClass == "Positive") %>% dplyr::summarise(sumPosSample = geometric.mean(posNorm), Sample = unique(Sample), Set = unique(Set)) %>% 
    ggplot(aes(x = Sample, y = sumPosSample)) + geom_point() + theme_bw()
```

## similarity between technical replicates (only using endogeneous mRNA trascripts)

```{r}
dat0 <- dataNorm %>% mutate(sample_filename = paste(Sample, fileName, sep = "_")) %>% 
  dplyr::select(sample_filename, posNorm, Name, CodeClass) %>% 
  filter(CodeClass == "Endogenous") %>% 
  spread(Name, posNorm)

eset0 <- as.matrix(dat0[, setdiff(colnames(dat0), c("sample_filename", "CodeClass"))])
result <- mixOmics::pca(eset0, scale = TRUE, center = TRUE)
group <- unlist(lapply(strsplit(dat0$sample_filename, "_"), function(i) i[1]))
group[group %in% names(table(group)[table(group) == 1])] <- "noTechnicalReplicate"
mixOmics::plotIndiv(result, group = group, ind.names = FALSE)

## turn data into matrix of expression
dat <- dataNorm %>% dplyr::select(Sample, posNorm, Name, CodeClass) %>% 
       group_by(Sample, Name, CodeClass) %>% 
       summarise_all(funs(mean))

## positive control normalized
eset_pos0 <- dat %>% filter(CodeClass == "Endogenous") %>% 
  spread(Name, posNorm)
eset_pos <- cpm(as.matrix(eset_pos0[, -c(1:2)]), log=T)
# eset_pos <- log2(as.matrix(eset_pos0[, -c(1:2)]))

rownames(eset_pos) <- eset_pos0$Sample


```

## house-keeping normalization

```{r}
## house-keeper normalization
hk <- filter(dat, CodeClass == "Housekeeping")
hkMean <- hk %>% 
  dplyr::group_by(Name) %>% 
  dplyr::summarise(meanCount = mean(posNorm))
hk$Name <- factor(as.character(hk$Name), levels = as.character(hkMean$Name))
hk$Sample <- factor(as.character(hk$Sample), levels = unique(as.character(hk$Sample))[order(as.numeric(gsub("E|V","", unique(as.character(hk$Sample)))))])

ggplot(hk, aes(x = as.numeric(Sample), y = posNorm, color = Name)) + geom_point() + geom_line() +
  scale_y_log10() + customTheme(sizeStripFont=15, xAngle=0, hjust = 0.5, vjust = 0.5, xSize=10, ySize=15, xAxisSize=15, yAxisSize=15) + ylab("Counts")
```

## determine whether hk normalization decreases overall CV of genes

```{r}
hkDat0 <- hk %>% dplyr::select(Name, Sample, posNorm) %>% spread(Name, posNorm)
hkDat <- as.matrix(hkDat0[, -1])
hkDat <- cpm(hkDat, log=T)
rownames(hkDat) <- as.character(hkDat0$Sample)

## Use geNorm to select house-keepers
res.BM <- selectHKs(hkDat, method = "geNorm", Symbols = colnames(hkDat), minNrHK = 2, log = FALSE)
hkRanking <- data.frame(Gene = as.character(res.BM$ranking), 
  Rank = 1:length(res.BM$ranking)) #,
#  Dataset = c(rep("UCSC genes", 3), rep("Trinity", 3)))

cvList <- list()
for(i in 3 : nrow(hkRanking)){
  genes <- as.character(hkRanking$Gene[1:i])
  
  ## House-keeping normalization factor
  hkNF <- dat %>% 
    filter(Name %in% genes) %>% 
    group_by(Sample) %>% dplyr::summarise(hkGeoMean = geometric.mean(posNorm)) %>% 
    mutate(hkNF = mean(hkGeoMean)/hkGeoMean)
  
  ## normalize data
  hkNorm <- full_join(dataNorm, hkNF, by = "Sample") %>% mutate(hkNorm = posNorm*hkNF)
  
  ## calculate CVs
  cvDat <- hkNorm %>% filter(CodeClass == "Endogenous") %>% 
    group_by(Name) %>% dplyr::summarise(cv = sd(hkNorm)/mean(hkNorm))
  cv <- cvDat$cv
  names(cv) <- cvDat$Name
  cvList[[i-2]] <- cv
}
names(cvList) <- paste0("hk", "1-", 3:nrow(hkRanking))
cvDatList <- as.data.frame(do.call(cbind, cvList))

## positive control normalization
cvDat <- dataNorm %>% filter(CodeClass == "Endogenous") %>% 
  group_by(Name) %>% dplyr::summarise(cv = sd(posNorm)/mean(posNorm))
cv <- cvDat$cv
names(cv) <- cvDat$Name
cvDatList$PosNorm <- cv
## raw data
cvDat <- asthmaDat %>% filter(CodeClass == "Endogenous") %>% 
  group_by(Name) %>% dplyr::summarise(cv = sd(Count)/mean(Count))
cv <- cvDat$cv
names(cv) <- cvDat$Name
cvDatList$Raw <- cv

boxplot(cvDatList, ylab = "coefficient of variation")
```

## plot normalize with top 3 house-keeping genes

```{r}
genes <- as.character(hkRanking$Gene[1:3])

## House-keeping normalization factor
hkNF <- dat %>% 
  filter(Name %in% genes) %>% 
  group_by(Sample) %>% dplyr::summarise(hkGeoMean = geometric.mean(posNorm)) %>% 
  mutate(hkNF = mean(hkGeoMean)/hkGeoMean)

## normalize data
hkNorm <- full_join(dat, hkNF, by = "Sample") %>% mutate(hkNorm = posNorm*hkNF) %>% 
          dplyr::filter(CodeClass == "Endogenous")

eset0 <- hkNorm %>% dplyr::select(Name, Sample, hkNorm) %>% 
        spread(Name, hkNorm)
eset_pos_hk <- as.matrix(eset0[, -1])
eset_pos_hk <- cpm(eset_pos_hk, log=T)
rownames(eset_pos_hk) <- as.character(eset0$Sample)

## do both mRNA datasets have the samples in the same order
all(rownames(eset_pos) == rownames(eset_pos_hk))
```

## Does the LAR correlate with AUC

```{r}
demo <- demo0[as.character(demo0$EorV_number) %in% rownames(eset_pos), ]
rownames(demo) <- as.character(demo$EorV_number)
demo <- demo[rownames(eset_pos), ]
date <- as.character(dataNorm$Date)
names(date) <- dataNorm$Sample
demo$RunDate <- factor(date[rownames(demo)])
demo$calculated_Response <- factor(demo$calculated_Response, c("ER", "DR"))
demo$SITE <- factor(demo$SITE)
demo$SEX <- as.character(demo$SEX)
all(rownames(demo) == rownames(eset_pos))
all(rownames(demo) == rownames(eset_pos_hk))

## compute area under the curve
require(pracma)
percentfev1 <- scale(t(demo[, c("BLFEV", "F10L", "F20L", "F30L","F45L", "F60L", "F90L", "F120L", "F180L", "F240L", "F300L", "F360L", "F420L")]),
  center = demo$BLFEV, scale = demo$BLFEV)
time <- c(0, 10, 20, 30, 45, 60, 90, 120, 180, 240, 300, 360, 420)
demo$AUC <- apply(percentfev1, 2, function(i){
  trapz(time, -i)
})
```

## Perform PCA and look for any batch effects

```{r}
vars <- setdiff(colnames(demo), c("SPONSOR", "DRUG", "ID", "NAME", "AIC_YMD", "rnaseqID_quebec",
  "Allergen", "UniqueID", "SubjInitials_mislabelled", "MST_LST_numbers", "Time",
  "Box", "concealedID", "concealed_uniqueID", "concealed_ID", "EorV_number", "RACE",
  "Mac_Response", "UniqueID", "Extracted", "AvailableRNAorPAXgene", "Cohort", "SubjInitials_mislabelled","MST_LST_numbers", "Response_usedInRNAseqBiomarkerAnalysis", "CorrectResponse", "rnaseq_coreSet_biomarkerAnalysis","ranseq_flipper", "rnaseq_BioReplicates", "PanCancer", "panCancer_biomarkerAnalysis","elements_noHBA2attenuated", "elements_noHBA2attenuated_biomarkerAnalysis", "biocrates", "biocrates_biomarkerAnalysis","elements_hba2attenuated", "elements_hba2attenuated_biomarkerAnalysis", "EorV_number", "Box", "SampleNumber_CIC", "concealed_uniqueID", "concealed_ID"))
compVar(demo = demo[rownames(eset_pos), ], eset = eset_pos, variables = vars, ncomp = 10)


compVar(demo = demo[rownames(eset_pos_hk), ], eset = eset_pos_hk, variables = vars, ncomp = 10)
```

## compile datasets

```{r}
pos_hk_demo1_dis <- eset_pos_hk[rownames(demo1_dis), ]
pos_hk_demo1_val <- eset_pos_hk[rownames(demo1_val), ]

## repeat challenges dataset
pos_hk_demo1_val_repeatDifferentResponse <- eset_pos_hk[rownames(demo1_val_repeatDifferentResponse), ]
pos_hk_demo1_val_repeatDR <- eset_pos_hk[rownames(demo1_val_repeatDR), ]

```

## compile all datasets

```{r}
## Positive+HK control normalized data
pos_hk_demo1_disList <- list(ucscGenes = pos_hk_demo1_dis[, ucscGenes], 
                          ucscGeneIso = pos_hk_demo1_dis[, ucscGeneIso], 
                          ensembl = pos_hk_demo1_dis[, ensembl], 
                          trinity = pos_hk_demo1_dis[, trinity], 
                          all = pos_hk_demo1_dis[, all])
pos_hk_demo1_valList <- list(ucscGenes = pos_hk_demo1_val[, ucscGenes], 
                          ucscGeneIso = pos_hk_demo1_val[, ucscGeneIso], 
                          ensembl = pos_hk_demo1_val[, ensembl], 
                          trinity = pos_hk_demo1_val[, trinity], 
                          all = pos_hk_demo1_val[, all])
lapply(pos_hk_demo1_disList, dim)
lapply(pos_hk_demo1_valList, dim)

## repeat challenges
# different response
pos_hk_demo1_val_repeatDifferentResponseList <- list(ucscGenes = pos_hk_demo1_val_repeatDifferentResponse[, ucscGenes], 
                          ucscGeneIso = pos_hk_demo1_val_repeatDifferentResponse[, ucscGeneIso], 
                          ensembl = pos_hk_demo1_val_repeatDifferentResponse[, ensembl], 
                          trinity = pos_hk_demo1_val_repeatDifferentResponse[, trinity], 
                          all = pos_hk_demo1_val_repeatDifferentResponse[, all])
# Reapted DRs
pos_hk_demo1_val_repeatDRList <- list(ucscGenes = pos_hk_demo1_val_repeatDR[, ucscGenes], 
                          ucscGeneIso = pos_hk_demo1_val_repeatDR[, ucscGeneIso], 
                          ensembl = pos_hk_demo1_val_repeatDR[, ensembl], 
                          trinity = pos_hk_demo1_val_repeatDR[, trinity], 
                          all = pos_hk_demo1_val_repeatDR[, all])

```

## add cbc data

```{r}
y.train_mrna <- factor(demo1_dis$calculated_Response, c("ER", "DR"))
names(y.train_mrna) <- rownames(demo1_dis)
y.test_mrna <- factor(demo1_val$calculated_Response, c("ER", "DR"))
names(y.test_mrna) <- rownames(demo1_val)
table(y.train_mrna); table(y.test_mrna)

clinical_train <- na.omit(demo1_dis[, c("BLFEV", "PC20", "Leukocyte_Counts.x10.9.", "Neu_percent","lym_percent","mono_percent","eos_percent","baso_percent")])
y.train_clinical <- y.train_mrna[rownames(clinical_train)]
clinical_test <- na.omit(demo1_val[, c("BLFEV", "PC20", "Leukocyte_Counts.x10.9.", "Neu_percent","lym_percent","mono_percent","eos_percent","baso_percent")])
y.test_clinical <- y.test_mrna[rownames(clinical_test)]
table(y.train_clinical); table(y.test_clinical)

## mRNA expression + cbc data
sapply(pos_hk_demo1_disList, function(i){all(rownames(i) == rownames(demo1_dis))})
sapply(pos_hk_demo1_valList, function(i){all(rownames(i) == rownames(demo1_val))})
## training datasets
pos_hk_demo1_disList$clinical <- as.matrix(clinical_train)
pos_hk_demo1_disList$ucscGenes_clinical <- as.matrix(cbind(clinical_train, pos_hk_demo1_disList$ucscGenes[rownames(clinical_train), ]))
pos_hk_demo1_disList$ucscGeneIso_clinical <- as.matrix(cbind(clinical_train, pos_hk_demo1_disList$ucscGeneIso[rownames(clinical_train), ]))
pos_hk_demo1_disList$ensembl_clinical <- as.matrix(cbind(clinical_train, pos_hk_demo1_disList$ensembl[rownames(clinical_train), ]))
pos_hk_demo1_disList$trinity_clinical <- as.matrix(cbind(clinical_train, pos_hk_demo1_disList$trinity[rownames(clinical_train), ]))
pos_hk_demo1_disList$all_clinical <- as.matrix(cbind(clinical_train, pos_hk_demo1_disList$all[rownames(clinical_train), ]))
## test datasets
pos_hk_demo1_valList$clinical <- as.matrix(clinical_test)
pos_hk_demo1_valList$ucscGenes_clinical <- as.matrix(cbind(clinical_test, pos_hk_demo1_valList$ucscGenes[rownames(clinical_test), ]))
pos_hk_demo1_valList$ucscGeneIso_clinical <- as.matrix(cbind(clinical_test, pos_hk_demo1_valList$ucscGeneIso[rownames(clinical_test), ]))
pos_hk_demo1_valList$ensembl_clinical <- as.matrix(cbind(clinical_test, pos_hk_demo1_valList$ensembl[rownames(clinical_test), ]))
pos_hk_demo1_valList$trinity_clinical <- as.matrix(cbind(clinical_test, pos_hk_demo1_valList$trinity[rownames(clinical_test), ]))
pos_hk_demo1_valList$all_clinical <- as.matrix(cbind(clinical_test, pos_hk_demo1_valList$all[rownames(clinical_test), ]))

y.train <- list(ucscGenes=y.train_mrna, ucscGeneIso=y.train_mrna, ensembl=y.train_mrna, 
                trinity=y.train_mrna, all=y.train_mrna, 
                clinical=y.train_clinical, ucscGenes_clinical=y.train_clinical, 
                ucscGeneIso_clinical=y.train_clinical, ensembl_clinical=y.train_clinical, 
                trinity_clinical=y.train_clinical, all_clinical=y.train_clinical)
y.test <- list(ucscGenes=y.test_mrna, ucscGeneIso=y.test_mrna, ensembl=y.test_mrna, 
                trinity=y.test_mrna, all=y.test_mrna, 
                clinical=y.test_clinical, ucscGenes_clinical=y.test_clinical, 
                ucscGeneIso_clinical=y.test_clinical, ensembl_clinical=y.test_clinical, 
                trinity_clinical=y.test_clinical, all_clinical=y.test_clinical)

```

## compile datasets

```{r}
save(demo1_dis=demo1_dis, demo1_val=demo1_val, y.train=y.train, y.test=y.test,
  pos_hk_demo1_disList=pos_hk_demo1_disList, pos_hk_demo1_valList=pos_hk_demo1_valList,
  demo1_val_repeatDifferentResponse=demo1_val_repeatDifferentResponse, 
  pos_hk_demo1_val_repeatDifferentResponseList=pos_hk_demo1_val_repeatDifferentResponseList,
  demo1_val_repeatDR=demo1_val_repeatDR, 
  pos_hk_demo1_val_repeatDRList=pos_hk_demo1_val_repeatDRList,
  file = meta_file_rnae_dir)

```

## SAVE

```{r}

meta_fileraw = get(load(paste0(meta_fileall_dir,".Rdata")))
meta_fileraw_id = paste0(meta_fileraw$id,"_",meta_fileraw$date,"_",meta_fileraw$time)

demo_ = demo0
demo_$NAME[grepl("WRF",demo_$NAME)] = "WRF"
demo_id = paste0(demo_$NAME,"_",demo_$AIC_YMD,"_",demo_$Time)
demo_ = demo_[match(meta_fileraw_id,demo_id),]
# demo_ = demo_[match(rownames(eset_pos_hk),rownames(demo_)),]
demo_ = demo_[!grepl("NA",rownames(demo_)),]

m <- eset_pos_hk[match(rownames(demo_),rownames(demo_)), ]
rownames(m) = demo_$NAME
save(m, file=paste0(feat_rnaelements_dir,".pre.Rdata"))
write.csv(m, file=paste0(feat_rnaelements_dir,".pre.csv"))


# meta_col
library(annotables)
grch38_dt <- merge(as.data.frame(grch38_tx2gene), as.data.frame(grch38), by="ensgene")
genes = sapply(strsplit(colnames(m),"_"), function(x) x[1])
genes[grepl("CECR1",genes)] = "ADA2"
genes[grepl("FAM65B",genes)] = "RIPOR2"
genes[grepl("CYTB",genes)] = "MT-CYB"
grep("unknown",genes)

gene_dt = grch38_dt[match(genes,grch38_dt$symbol),]
which(is.na(gene_dt$symbol)) # "noNamelncRNA" "unknown2a"    "unknown2"     "unknown1"     "CECR1"        "unknown3"     "FAM65B" "CYTB" 
gene_dt[is.na(gene_dt$symbol),c("start","end","chr")] = 0
gene_dt[is.na(gene_dt$symbol),c("symbol")] = genes[is.na(gene_dt$symbol)]

gene_dt$id = colnames(m)
save(gene_dt, file=paste0(meta_col_rnaelements_dir,".Rdata"))
write.csv(gene_dt, file=paste0(meta_col_rnaelements_dir,".csv"))
```



```
