==> 01a_analysis_metab-reformat.Rmd <==
---
title: "Discovery metabolite datasets"
output:
  html_document:
    theme: paper
    highlight: zenburn
    number_sections: true
    toc: true
    toc_float: 
      collapsed: true
      smooth_scroll: true

---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, tidy = TRUE, warning = TRUE, message = FALSE)
```

### specify path and load libraries

```{r}
## logistics
root = "~/projects/asthma" # root directory, used for _dirs.R
source(paste0(root, "/src/_dirs.R"))
source(paste0(root, "/src/_func.R"))
source(paste0(root, "/src/_func-asthma.R"))
libr(append(pkgs(),c("tidyverse","gplots","mixOmics",
                     "igraph","NMF","ggrepel","amritr")))
```

# Metabolomics dataset 
  * dataset size: 112 samples x 163 metabolites
  * 16 plasama samples (plate 1) (didn't extract data for biocrates control samples)
  * 88 plasma samples + 8 biocrates controls (1 blank, 3PBS samples, 1 standard and 3 quality control samples)
  
```{r}
## demographics
demo <- readRDS(meta_file_rnaseqa_dir)
rownames(demo) <- paste(demo$concealedID, demo$Time, sep = "_")

##===================================
#
# 88 plasma samples
#
##===================================
plate.96wells <- read.csv(asthma_discov_dir, stringsAsFactors=FALSE)
# colnames(plate.96wells) = plate.96wells[1,]
# plate.96wells = plate.96wells[-1,]
plateEset0 <- plate.96wells %>% slice(-c(1:14)) %>% dplyr::select(C0:H1) %>% apply(., 2, as.numeric)
lodSOP0 <- dplyr::select(plate.96wells[14, ], C0:H1) %>% as.numeric
plateSamples <- with(plate.96wells[-c(1:14),], paste(concealedID, Time, sep = "_"))

meta = data.frame(time=plate.96wells$Time, 
                  date=gsub("[.]", "-", plate.96wells$Collection.Date))
meta$id = plate.96wells$Sample.Identification
meta$id[grepl("WRF",meta$id)] = "WRF"

names(lodSOP0) <- colnames(plateEset0)
all(colnames(plateEset0) == names(lodSOP0))

## metabolite annotation
metaboliteType <- plate.96wells[2,] %>% dplyr::select(C0:H1) %>% as.character %>% factor
names(metaboliteType) <- colnames(plateEset0)

## Since the colnames don't match, reorder columns of the plate metabolite expression dataset
lodSOP <- lodSOP0[colnames(plateEset0)]
all(colnames(plateEset0) == colnames(lodSOP))
all(colnames(plateEset0) == colnames(metaboliteType))

## label samples
sampleGroup <- plateSamples
sampleGroup[grep("HLI", sampleGroup)] <- "Plasma Samples"
sampleGroup[grep("Pre", sampleGroup)] <- "Plasma Samples"
sampleGroup[grep("Post", sampleGroup)] <- "Plasma Samples"
sampleGroup[which(!(sampleGroup %in% "Plasma Samples"))] <- "Biocrates"

## remove biocrates samples
metEset <- plateEset0[sampleGroup != "Biocrates", ]
rownames(metEset) <- plateSamples[sampleGroup != "Biocrates"]
metDemo <- demo[rownames(metEset), ]
all(rownames(metEset) == paste(metDemo$concealedID, metDemo$Time, sep="_"))

```

## plate 2 - metabolite expression

```{r}
metDemo.plate <- subset(metDemo, !(UniqueID %in% c("L_ST_026_R1", "L_ST_005_A")))
metEset.plate <- as.matrix(metEset)[paste(metDemo.plate$concealedID, metDemo.plate$Time, sep="_"),] %>% 
  as.data.frame %>%
  mutate(concealedID = metDemo.plate$concealedID, Time = metDemo.plate$Time) %>% 
  group_by(concealedID, Time) %>% arrange(concealedID, Time) %>% 
  summarise_each(funs(mean))


eset0 <- as.matrix(metEset.plate[, names(lodSOP)])
eset0 <- eset0[, apply(eset0, 2, mean) > lodSOP]
rownames(eset0) <- paste(metEset.plate$concealedID, metEset.plate$Time, sep="_")
demo <- metDemo[rownames(eset0), ]

eset <- 1000*eset0
eset[eset < 1] <- 1
eset <- log2(eset)
all(rownames(eset) == rownames(demo))

## biological replicates were removed L_ST_005_A and L_ST_026_R1
plot(table(as.character(demo$NAME)))   ##  pre and post

## Convert metabolites to parent metabolite modules
metNamesDat <- read.csv(metab_names_dir, row.names = 1)
metNames <- as.character(metNamesDat$parent_metabolite)
names(metNames) <- rownames(metNamesDat)
metNames <- metNames[colnames(eset)]
all(colnames(eset) == names(metNames))

```

```{r}
demoAsthma <- subset(demo, givenResponse != "Control")
demoAsthma$givenResponse <- factor(demoAsthma$givenResponse, levels = c("ER", "DR"))
demoAsthma$BMI <- demoAsthma$Wt..Kg./(demoAsthma$HT.cm./100)^2
demoAsthma$SEX <- droplevels(demoAsthma$SEX)

demoAsthma <- demoAsthma[!is.na(demoAsthma$BMI), ]
esetAsthma <- eset[rownames(demoAsthma), ]
all(rownames(demoAsthma) == rownames(esetAsthma))

## pre-challenge
demoAsthmaPre <- subset(demoAsthma, Time == "Pre")
esetAsthmaPre_ = esetAsthmaPre <- esetAsthma[rownames(demoAsthmaPre), ]


## post-challenge (scaled to pre-challenge)
demoAsthmaPost = demoAsthmaDiff <- subset(demoAsthma, Time == "Post")
all(demoAsthmaPre$NAME == demoAsthmaDiff$NAME)
esetAsthmaPost_ <- esetAsthmaPost <- esetAsthma[rownames(demoAsthmaDiff), ]
esetAsthmaDiff_ <- esetAsthmaDiff <- esetAsthmaPost - esetAsthmaPre

```

## save pre and post data matrix!

```{r}

namepre = as.character(demoAsthmaPre$NAME)
namepre[grepl("WRF",namepre)] = "WRF"

rownames(esetAsthmaPre_) = rownames(esetAsthmaPost_) = rownames(esetAsthmaDiff_) = namepre

# save(esetAsthmaPre_, file=paste0(feat_metab_dir,".pre.Rdata"))
# save(esetAsthmaPost_, file=paste0(feat_metab_dir,".post.Rdata"))
# save(esetAsthmaDiff_, file=paste0(feat_metab_dir,".diff.Rdata"))
# write.csv(esetAsthmaPre_, file=paste0(feat_metab_dir,".pre.csv"))
# write.csv(esetAsthmaPost_, file=paste0(feat_metab_dir,".post.csv"))
# write.csv(esetAsthmaDiff_, file=paste0(feat_metab_dir,".diff.csv"))

```

## plot FEV

```{r}
## FEV1 plot
demo.core.fev1 <- demoAsthma[, c("BLFEV", "F10L","F20L","F30L",
                                    "F45L","F60L", "F90L","F120L",
                                    "F180L","F240L","F300L","F360L","F420L")]
demo.core.fev1.scale <- 100*scale(t(demo.core.fev1), center=demoAsthma$BLFEV, scale=demoAsthma$BLFEV)

fev1.list <- list()
for(i in 1:ncol(demo.core.fev1.scale)){
  fev1.list[[i]] <- as.matrix(demo.core.fev1.scale[,i])
}
fev1.rbind <- do.call(rbind, fev1.list)

fev1.time <- demo.core.fev1.scale %>% as.data.frame %>% 
  mutate(Time = rownames(.)) %>% gather(Subj, Fev1, -Time)
fev1.time$Time[fev1.time$Time == "BLFEV"] <- 0
fev1.time$Time <- fev1.time$Time %>% gsub("F", "", .) %>% gsub("L", "", .) %>% as.numeric 
fev1.time$Time <- fev1.time$Time/60
fev1.time$Response <- demoAsthma[fev1.time$Subj, "calculated_Response"]
fev1.time$Response[fev1.time$Response == "ER"] <- "Early Responders"
fev1.time$Response[fev1.time$Response == "DR"] <- "Dual Responders"

addLabels <- data.frame(Fev1 = c(-20, -18),
  Time = c(2.5, 5), Res = c("EAR", "LAR"),
  Response = c("Early Responders", "Early Responders"), Col = c("Early Responders", "Early Responders"), Subj = c("Subj110_Pre", "Subj110_Pre"))

pdf(meta_metab_1a_dir, width = 8, height = 4.5)
ggplot(fev1.time, aes(x = Time, y = Fev1, group=Subj, colour=Response)) + geom_line() +
  stat_smooth(aes(group = 1)) + stat_summary(aes(group = 1),
                                             geom = "point", fun.y = mean, shape = 17, size = 3) + facet_grid(. ~ Response) +
  scale_y_continuous(expression('Percent drop in'~ FEV[1])) + #theme_bw() +
  scale_x_continuous(expression('Time (hours)')) +
  theme(axis.text.y = element_text(size = 15, hjust = 1)) + theme(axis.text.x = element_text(size = 15, hjust = 0.5))+
  theme(axis.title.x=element_text(size = 15)) + theme(axis.title.y=element_text(size = 15,angle = 90))+ 
  theme(plot.background = element_rect()) +   theme_bw() + 
  theme(strip.background = element_rect(colour = "black", fill = "white",
                                        size = 1), strip.text.x = element_text(size=20)) +
        theme(legend.position="none") + 
  geom_segment(aes(x = -0.08, y = -60, xend = -0.08, yend = -55), arrow = arrow(length = unit(0.3, "cm")), col = "black") +
  geom_segment(aes(x = 2, y = -60, xend = 2, yend = -55), arrow = arrow(length = unit(0.3, "cm")), col = "black") +
  annotate("text", x=1, y=-58, label="Blood draw") +
  geom_label(data = addLabels, aes(label = Res)) +
  geom_segment(aes(x = 0, y = -20, xend = 2, yend = -20), color = "gray", linetype = "dashed") +
  geom_segment(aes(x = 3, y = -15, xend = 7, yend = -15), color = "gray", linetype = "dashed") + 
  scale_color_manual(values=c("#F8766D", "#619CFF"))
dev.off()

```

# demographics

```{r}
variables <- c("AGE", "BMI", "EAR", "LAR", "PRFEV", "AIS")
descriptiveStat(demo = demoAsthmaPre, groups = "calculated_Response", variables = variables)

table(demoAsthmaPre$SEX, demoAsthmaPre$calculated_Response)
table(demoAsthmaPre$Allergen_cleanLabel, demoAsthmaPre$calculated_Response)
table(demoAsthmaPre$SITE, demoAsthmaPre$calculated_Response)

## Control
subset(demo, givenResponse == "Control" & Time == "Pre")
mean(subset(demo, givenResponse == "Control" & Time == "Pre")$AGE)
sd(subset(demo, givenResponse == "Control" & Time == "Pre")$AGE)

```

## ER vs. DR

```{r}
#Pre
design <- model.matrix(~givenResponse+AGE+SEX+BMI, data = demoAsthmaPre)
fit <- eBayes(lmFit(t(esetAsthmaPre), design, method = "robust", maxit = 100), robust = TRUE, trend = TRUE)
top.pre <- topTable(fit, coef = 2, adjust.method = "BH", n = Inf)

#Interaction
design <- model.matrix(~givenResponse+AGE+SEX+BMI, data = demoAsthmaDiff)
fit <- eBayes(lmFit(t(esetAsthmaDiff), design, method = "robust", maxit = 100), robust = TRUE, trend = TRUE)
top.int <- topTable(fit, coef = 2, adjust.method = "BH", n = Inf)

```

## Figure 2 - p-value histogram

```{r}
jpeg(meta_metab_1b_dir, width = 600, height = 400)
data.frame(Interaction = top.int$P.Value,
  Pre = top.pre$P.Value) %>% 
  gather(Comparison, p.value) %>% 
  ggplot(aes(x = p.value, fill = Comparison)) +
  geom_histogram(bins=20, alpha=.9, position="identity") +
  customTheme(sizeStripFont = 10, xAngle = 0, hjust = 0.5,
    vjust = 0.5, xSize = 10, ySize = 10, xAxisSize = 10,
    yAxisSize = 10) + ylim(c(0, 17)) +
  geom_hline(yintercept = 0.05*nrow(top.pre), linetype = "dashed") +
  annotate("text", x = 0.84, y = 7.5, parse = TRUE, label = as.character(expression("Expected distribution if"~ H[0]~"is true"))) + ylab("Frequency") + xlab("P-values") +
  ggtitle("P-value histograms") + 
  theme(legend.position = c(0.9, 0.9))
dev.off()



```

## Figure 3

```{r}
top.pre$logPval <- -log10(top.pre$P.Value)
top.pre$Metabolite <- rownames(top.pre)
top.pre$Type <- droplevels(metaboliteType[top.pre$Metabolite])

thres <- top.pre$logPval[top.pre$adj.P.Val < 0.3]
cbPalette <- c("#66C2A5","#E78AC3", "#FC8D62", "#8DA0CB",  "#A6D854")
deGen <- ggplot(top.pre, aes(x = logFC, y = logPval, color = Type)) + geom_point(size = 2) +
  geom_hline(aes(yintercept = thres[length(thres)]), lty = 2, col = "red") +
  geom_vline(aes(xintercept = 0), lty = 2, color = "gray") +
  geom_text_repel(data=filter(top.pre, logPval >= thres[length(thres)]), aes(label=Metabolite)) + customTheme(sizeStripFont = 10, xAngle = 0, hjust = 0.5, vjust = 0.5, xSize = 15, ySize = 15, xAxisSize = 15, yAxisSize = 15) +
  scale_x_continuous(expression("log"[2]~"Fold-change"), limits = c(-0.8, 0.5)) +
  scale_y_continuous(expression("-log"[10]~"P-value")) +
  annotate("text", x = -0.5, y = 1.5, label = "BH-FDR=30%", col = "red") +
  ggtitle("Differential Metabolite expression") + theme(legend.position = c(0.19, 0.8))+
  scale_colour_manual(values=cbPalette)

pdf(meta_metab_1c_dir, width = 6, height = 6)
deGen
dev.off()


```

## Supplement figure ER-DR-Control

```{r}
topMets <- rownames(top.pre)[top.pre$adj.P.Val < 0.3]
demoPre <- subset(demo, Time == "Pre")
esetPre <- eset[rownames(demoPre), ]
design <- model.matrix(~givenResponse+0, data = demoPre)
contMatrix <- makeContrasts(Control.ER=givenResponseER-givenResponseControl, Control.DR=givenResponseDR-givenResponseControl,
  levels = design)
fit <- lmFit(t(esetPre[, topMets]), design, robust = TRUE)
fit <- eBayes(contrasts.fit(fit, contMatrix), robust = TRUE, trend = TRUE)
allComp <- decideTests(fit, method = "global", p.value = 0.3)
summary(allComp)
sigMet.Controls.er.dr <- rownames(as.data.frame(allComp)[as.data.frame(allComp)[,1] != 0, ])

pdf(meta_metab_1e_dir, width = 5, height = 5)
esetPre[, sigMet.Controls.er.dr] %>% 
  as.data.frame %>% 
  mutate(Group = demoPre$givenResponse) %>% 
  gather(Metabolite, Exp, -Group) %>% 
  ggplot(aes(x = Group, y = Exp, fill = Group)) +
  geom_boxplot() +
  facet_wrap(~Metabolite, scales = "free") +
  customTheme(sizeStripFont = 10, xAngle = 0,
    hjust = 0.5, vjust = 0.5, xSize = 10, ySize = 10,
    xAxisSize = 10, yAxisSize = 10) + ylab( expression("Concentration  ("~log[2]~"mM)")) + 
  scale_fill_manual(values=c("#00BA38", "#F8766D", "#619CFF"))
dev.off()

```


## correlatoin between cell-types

```{r}
cbc <- demoAsthmaPre[, c("Neu_percent","lym_percent",
  "mono_percent","eos_percent", "baso_percent")]
colnames(cbc) <- c("Neutrophils", "Lymphocytes",
  "Monocytes", "Eosinophils", "Basophils")
cbc <- na.omit(cbc)

boxplot(cbc) 
## remove outliers
outliers <- mahalanobis(cbc, center = colMeans(cbc), cov = cov(cbc))
pcaCBC <- pca(cbc, scale = TRUE, center = TRUE)
plotIndiv(pcaCBC)

cbc <- cbc[!(rownames(cbc) %in% names(outliers[outliers > 20])), ]
pcaCBC <- pca(cbc, scale = TRUE, center = TRUE)
plotIndiv(pcaCBC)

selectMets <- esetAsthmaPre[rownames(cbc), topMets]
selectDemo <- demoAsthmaPre[rownames(selectMets), ]

apply(cbc, 2, function(i){
  summary(lm(i ~ selectDemo$givenResponse))$coef[2,"Pr(>|t|)"]
})

fcDr <- cor(cbc[selectDemo$givenResponse == "DR",], selectMets[selectDemo$givenResponse == "DR", topMets], use = "na.or.complete", method = "spearman")
fcEr <- cor(cbc[selectDemo$givenResponse == "ER",], selectMets[selectDemo$givenResponse == "ER", topMets], use = "na.or.complete", method = "spearman")
fcDr[fcDr < 0] <- 0
fcEr[fcEr < 0] <- 0
fc = fcDr-fcEr
heatmap.2(fc, trace = "none", margins = c(10, 10))

metLabels <- factor(as.character(metaboliteType[colnames(fc)]))
NMF::aheatmap(fc, annCol = list(Type = metLabels),
  annColors = "Set2",
  filename = meta_metab_1d_dir, 
  border = TRUE, width = 7, height = 7,
  main = "")


pdf(meta_metab_1da_dir, width = 4, height = 4)
data.frame(PC.ae.C42.2 = selectMets[, "PC.ae.C42.2"]) %>% 
  mutate(Basophils = cbc$Basophils,
         Monocytes = cbc$Monocytes,
         Response = relevel(selectDemo$givenResponse, ref = "DR")) %>% 
  gather(Cell, Prop, -c(PC.ae.C42.2, Response)) %>% 
  mutate(Cell = factor(Cell, levels = c("Monocytes", "Basophils"))) %>% 
  ggplot(aes(x = Prop, y = PC.ae.C42.2)) +
  geom_point() + facet_wrap(Cell~Response, scales = "free") +
  stat_smooth(method = "lm") + customTheme(sizeStripFont = 10, xAngle = 30, hjust = 1, vjust = 1, xSize = 10, 
    ySize = 10, xAxisSize = 10, yAxisSize = 10) + ylab( expression("Concentration of PC.ae.C42.2 ("~log[2]~"mM)")) +
  xlab("Proportion of cell-types")
dev.off()


```
