==> 01a_analysis_metab.Rmd <==
---
title: "Metabolite biomarkers of the late phase asthmatic response"
output: 
  html_document:
    theme: paper
    highlight: zenburn
    number_sections: true
    toc: true
    toc_float: 
      collapsed: true
      smooth_scroll: true

vignette: >
  %\VignetteIndexEntry{"Introduction to IHW"}
  %\VignetteEngine{knitr::rmarkdown}
  %\usepackage[utf8]{inputenc}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, cache = TRUE, warning = FALSE, message = FALSE)
```
\newpage

# Metabolomics dataset 
  * 88 plasma samples + 8 biocrates controls (1 blank, 3PBS samples, 1 standard and 3 quality control samples)

```{r import datasets, message=FALSE, warning=FALSE}
## logistics
root = "~/projects/asthma" # root directory, used for _dirs.R
source(paste0(root, "/src/_dirs.R"))
source(paste0(root, "/src/_func.R"))
source(paste0(root, "/src/_func-asthma.R"))
libr(append(pkgs(),c("knitr","grid","gridExtra", "ggbiplot", "rafalib",
                     "GGally","NMF","ROCR","sva","mixOmics","iDINGO", "glmnet",
                     "RColorBrewer","igraph","pROC","OptimalCutpoints")))
# jPaste <- function(...) paste(..., sep = "")

## demographics
demo <- readRDS(meta_file_rnaseqa_dir)
rownames(demo) <- paste(demo$concealedID, demo$Time, sep = "_")
##===================================
#
# Plate 2: 88 plasma samples
#
##===================================
plate2.96wells <- read.csv(asthma_discov_dir, stringsAsFactors=FALSE)
plate2Eset0 <- plate2.96wells %>% slice(-c(1:14)) %>% dplyr::select(C0:H1) %>% apply(., 2, as.numeric)
lodSOP0 <- dplyr::select(plate2.96wells[14, ], C0:H1) %>% as.numeric
plate2Samples <- with(plate2.96wells[-c(1:14),], paste(concealedID, Time, sep = "_"))
names(lodSOP0) <- colnames(plate2Eset0)
all(colnames(plate2Eset0) == names(lodSOP0))

metaboliteType <- plate2.96wells[2,] %>% dplyr::select(C0:H1) %>% as.character %>% factor
names(metaboliteType) <- colnames(plate2.96wells)[-c(1:22)]


```

# plot correlation between samples

```{r}
subj <- unlist(lapply(strsplit(plate2Samples, "_"), function(i) i[1]))
time <- unlist(lapply(strsplit(plate2Samples, "_"), function(i) i[2]))
plate2Eset <- plate2Eset0 %>% as.data.frame %>% mutate(Subj = subj, Time = time) %>% 
  group_by(Subj, Time) %>% summarise_each(funs(mean))

metEset20 <- as.matrix(plate2Eset[, -c(1:2)])
rownames(metEset20) <- paste(plate2Eset$Subj, plate2Eset$Time, sep = "_")
metEset2 <- metEset20[rownames(metEset20) %in% rownames(demo), ]
metDemo2 <- demo[rownames(metEset2), ]
dim(metEset2);
table(metDemo2$calculated_Response, metDemo2$Time)

```

##

```{r}
metDemo2Asthma <- metDemo2[metDemo2$calculated_Response != "Control", ]
metDemo2AsthmaPre <- subset(metDemo2Asthma, Time == "Pre")
metDemo2AsthmaPost <- subset(metDemo2Asthma, Time == "Post")

## FEV1 plot
demo.core.fev1 <- metDemo2AsthmaPre[, c("BLFEV", "F10L","F20L","F30L",
                                    "F45L","F60L", "F90L","F120L",
                                    "F180L","F240L","F300L","F360L","F420L")]
demo.core.fev1.scale <- 100*scale(t(demo.core.fev1), center=metDemo2AsthmaPre$BLFEV, scale=metDemo2AsthmaPre$BLFEV)

fev1.list <- list()
for(i in 1:ncol(demo.core.fev1.scale)){
  fev1.list[[i]] <- as.matrix(demo.core.fev1.scale[,i])
}
fev1.rbind <- do.call(rbind, fev1.list)

fev1.time <- as.data.frame(demo.core.fev1.scale)
fev1.time.melt <- melt(fev1.time)
fev1.dat <- data.frame(fev1=as.numeric(fev1.time.melt$value))
fev1.dat$Time <- rep(round(c(0,10,20,30,45,60,90,120,180,240,300,360,420)/60,2), ncol(demo.core.fev1.scale))
fev1.dat$Subj <- factor(rep(colnames(demo.core.fev1.scale), each=nrow(demo.core.fev1.scale)))
res <- as.character(metDemo2AsthmaPre$calculated_Response)
res[res == "ER"] <- "Early Responders"
res[res == "DR"] <- "Dual Responders"
fev1.dat$Response <- factor(rep(res, each=nrow(demo.core.fev1.scale)), levels=c("Early Responders","Dual Responders"))
col.subj <- factor(res, levels=c("Early Responders","Dual Responders"))
#col.subj[which(demo.core.pre[colnames(demo.core.fev1.scale), "Participant.ID"] == "L_ST_005")] <- "DR2"
#col.subj[which(demo.core.pre[colnames(demo.core.fev1.scale), "Participant.ID"] == "L_ST_026")] <- "ER2"
fev1.dat$Col <- factor(rep(col.subj, each=nrow(demo.core.fev1.scale)))

addLabels <- data.frame(fev1 = c(-20, -18),
  Time = c(2.5, 5), Res = c("EAR", "LAR"),
  Response = c("Early Responders", "Early Responders"), Col = c("Early Responders", "Early Responders"), Subj = c("Subj110_Pre", "Subj110_Pre"))

pdf(meta_metab_1a_dir, width = 8, height = 5)
ggplot(fev1.dat, aes(x = Time, y = fev1, group=Subj, colour=Col)) + geom_line() +
  stat_smooth(aes(group = 1)) + stat_summary(aes(group = 1),
                                             geom = "point", fun.y = mean, shape = 17, size = 3) + facet_grid(. ~ Response) +
  scale_y_continuous(expression('Percent drop in'~ FEV[1])) + #theme_bw() +
  scale_x_continuous(expression('Time (hours)')) +
  theme(axis.text.y = element_text(size = 15, hjust = 1)) + theme(axis.text.x = element_text(size = 15, hjust = 0.5))+
  theme(axis.title.x=element_text(size = 15)) + theme(axis.title.y=element_text(size = 15,angle = 90))+ 
  theme(plot.background = element_rect()) +   theme_bw() + 
  theme(strip.background = element_rect(colour = "black", fill = "white",
                                        size = 1), strip.text.x = element_text(size=20)) +
        theme(legend.position="none") + 
  geom_segment(aes(x = -0.08, y = -60, xend = -0.08, yend = -55), arrow = arrow(length = unit(0.3, "cm")), col = "black") +
  geom_segment(aes(x = 2, y = -60, xend = 2, yend = -55), arrow = arrow(length = unit(0.3, "cm")), col = "black") +
  annotate("text", x=1, y=-58, label="Blood draw") +
  geom_label(data = addLabels, aes(label = Res)) +
  geom_segment(aes(x = 0, y = -20, xend = 2, yend = -20), color = "gray", linetype = "dashed") +
  geom_segment(aes(x = 3, y = -15, xend = 7, yend = -15), color = "gray", linetype = "dashed")
dev.off()

```

# demographics

```{r}
metDemo2AsthmaPre$BMI <- metDemo2AsthmaPre$Wt..Kg./(metDemo2AsthmaPre$HT.cm./100)^2
variables <- c("AGE", "BMI", "EAR", "LAR", "BLFEV", "AIS")
descriptiveStat(demo = metDemo2AsthmaPre, groups = "calculated_Response", variables = variables)

table(metDemo2AsthmaPre$SEX, metDemo2AsthmaPre$calculated_Response)
table(metDemo2AsthmaPre$Allergen_cleanLabel, metDemo2AsthmaPre$calculated_Response)
table(metDemo2AsthmaPre$SITE, metDemo2AsthmaPre$calculated_Response)

## Control
subset(metDemo2, calculated_Response == "Control" & Time == "Pre")
mean(subset(metDemo2, calculated_Response == "Control" & Time == "Pre")$AGE)
sd(subset(metDemo2, calculated_Response == "Control" & Time == "Pre")$AGE)
```

## remove metabolite with average expression lower than the lod

```{r}
all(rownames(lodSOP0) == colnames(metEset2))

metEset2 <- metEset2[, apply(metEset2, 2, mean) > lodSOP0]
```

## Build a classifer - biomarker panel

```{r}
gp <- factor(metDemo2AsthmaPre$calculated_Response, levels = c("ER", "DR"))

X <- metEset2[rownames(metDemo2AsthmaPre), ]
pb <- unlist(lapply(1:nrow(X), function(i){
  fit <- glmnet(X[-i, ], gp[-i], family = "binomial", alpha = 0.9)
cv.fit <- cv.glmnet(X[-i, ], gp[-i], family = "binomial")
predict(fit, newx = X[i, , drop = FALSE], s = cv.fit$lambda.min, type = "response")
}))

par(mfrow = c(1, 2))
plot(pb, col = gp,
 ylab = "P(Response = DR)", xlab = "36 subjects")
legend("topright", c("ER", "DR"), 
  col = 1:2, pch = 19, bty = "n")
plot(pb ~ gp, col = 1:2,
 ylab = "P(Response = DR)", xlab = "36 subjects")
legend("topright", c("ER", "DR"), 
  col = 1:2, pch = 19, bty = "n")

par(mfrow = c(1, 1))
pROC::roc(gp ~ pb, levels = c("ER", "DR"), plot = TRUE)
text(x = 0.2, y = 0.2, labels = "AUC = 87%")

pROC::roc(gp ~ pb, levels = c("ER", "DR"), plot = TRUE, direction = "<")


```

## leave-pair-out cross validation

```{r}
pairsSubj <- expand.grid(which(gp=="ER"), which(gp=="DR"))
dim(pairsSubj)

pb2 <- apply(pairsSubj, 1, function(i){
  heldOutSet <- as.numeric(i)
    fit <- glmnet(X[-heldOutSet, ], gp[-heldOutSet], family = "binomial", alpha = 0.9)
    cv.fit <- cv.glmnet(X[-heldOutSet, ], gp[-heldOutSet], family = "binomial")
    predict(fit, newx = X[heldOutSet, , drop = FALSE], s = cv.fit$lambda.min, type = "response")
})

gp2 <- as.character(gp)
predProb <- c(pb2[1,], pb2[1,])
trueLabels <- factor(c(gp2[pairsSubj[,1]], gp2[pairsSubj[,2]]), levels = c("ER", "DR"))
par(mfrow = c(1, 2))
boxplot(predProb ~ trueLabels, col = 1:2, ylim = c(0.5,1),
 ylab = "P(Response = DR)", xlab = "36 subjects")
legend("topright", c("ER", "DR"), 
  col = 1:2, pch = 19, bty = "n")

pROC::roc(trueLabels, predProb, levels = c("ER", "DR"), plot = TRUE)
text(x = 0.2, y = 0.2, labels = "AUC = 50%")




```
