==> 01a_analysis_rnaelements.Rmd <==
---
title: "latePhaseBiomarkers_DataAnalysis"
output:
  # pdf_document:
  #   toc: yes
  html_document:
    theme: paper
    highlight: zenburn
    number_sections: true
    toc: true
    toc_float: 
      collapsed: true
      smooth_scroll: true
---

```{r, echo = FALSE, results='hide', message=FALSE, warning=FALSE}
knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)

## logistics
root = "~/projects/asthma" # root directory, used for _dirs.R
source(paste0(root, "/src/_dirs.R"))
source(paste0(root, "/src/_func.R"))
source(paste0(root, "/src/_func-asthma.R"))
source(paste0(root, "/src/_func-classifiers.R"))
libr(append(pkgs(),c("mixOmics","tidyverse","amritr","nlme","elasticnet")))

load(meta_file_rnae_dir)
```

## plot fev1 profiles of discovery and validation dataset

```{r}
demoPre <- rbind(demo1_dis, demo1_val)
demoPre$Set <- rep(c("Discovery cohort", "Validation cohort"), c(nrow(demo1_dis), nrow(demo1_val)))
  
demo.core.fev1 <- demoPre[, c("BLFEV", "F10L","F20L","F30L",
                                    "F45L","F60L", "F90L","F120L",
                                    "F180L","F240L","F300L","F360L","F420L")]
demo.core.fev1.scale <- 100*scale(t(demo.core.fev1), center=demo.core.fev1$BLFEV, scale=demo.core.fev1$BLFEV)

fev1Dat <- demo.core.fev1.scale %>% as.data.frame %>% 
  mutate(time = c(0, as.numeric(gsub("F|L", "", rownames(demo.core.fev1.scale)[-1])))/60) %>% 
  gather(Subj, dropFEV1, -time)
pdf(paste0(rnaelements_dir, "/results/Figure4a.pdf"), width = 15, height = 6.5)
fev1Dat %>% mutate(Set = factor(demoPre[fev1Dat$Subj, "Set"], c("Discovery cohort", "Validation cohort")), 
         Response = factor(demoPre[fev1Dat$Subj, "calculated_Response"], levels = c("ER", "DR")),
         subj = demoPre[fev1Dat$Subj, "concealed_uniqueID"]) %>% 
  ggplot(aes(x = time, y = dropFEV1, group = subj, color = Response)) +
  geom_segment(aes(x = 3, y = -15, xend = 7, yend = -15), col = "black", linetype = "dashed") +
  geom_point(size=3) +
  geom_line(size=1) + 
  stat_smooth(aes(group = Response)) + stat_summary(aes(group = Response), geom = "point", fun.y = mean, shape = 17, size = 3) +
  facet_wrap(~Set) +
  customTheme(sizeStripFont = 25, xAngle = 0, hjust = 0.5, vjust = 0.5, xSize = 20, ySize = 20, xAxisSize = 20, yAxisSize = 20) +
  geom_segment(aes(x = -0.08, y = -60, xend = -0.08, yend = -55), arrow = arrow(length = unit(0.3, "cm")), col = "black") +
  annotate("text", x=1.7, y=-58, label="Blood draw", cex = 7) +
  scale_y_continuous(expression('Percent drop in '~ FEV[1])) + xlab("Time (hours)") +
  theme(legend.text = element_text(size = 20)) +
  geom_hline(yintercept = 0, linetype = "dashed", col = "gray")
dev.off()

```

## demographics

```{r}
nrow(demo1_dis); nrow(demo1_val)
table(demo1_dis$calculated_Response); table(demo1_val$calculated_Response)
table(demo1_val$SEX, demo1_val$calculated_Response)
11/24 # female (dr)
5/9 # female (er)
table(as.character(demo1_val$Allergen_cleanLabel), demo1_val$calculated_Response)
table(demo1_val$SITE, demo1_val$calculated_Response)

## Weight
fit <- lm(demo1_val$Wt..Kg. ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # p=0.51
## height
fit <- lm(demo1_val$HT.cm. ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) #pt = 0.43
## Age 
fit <- lm(demo1_val$AGE ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## not normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = demo1_val$AGE[demo1_val$calculated_Response == "ER"], 
            y = demo1_val$AGE[demo1_val$calculated_Response == "DR"])  # pw=0.35
## BLFEV 
fit <- lm(demo1_val$BLFEV ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) #p = 0.14
## PRFEV
fit <- lm(demo1_val$PRFEV ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) #p = 0.98
## EAR
fit <- lm(demo1_val$EAR ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) #p = 0.21
## LAR
fit <- lm(demo1_val$LAR ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## not normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = demo1_val$LAR[demo1_val$calculated_Response == "ER"], 
            y = demo1_val$LAR[demo1_val$calculated_Response == "DR"])  # p=1.388e-05
## PC20
fit <- lm(demo1_val$PC20 ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## not normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = demo1_val$PC20[demo1_val$calculated_Response == "ER"], 
            y = demo1_val$PC20[demo1_val$calculated_Response == "DR"])  # pw=0.49
## AIS
fit <- lm(demo1_val$AIS ~ demo1_val$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # Pt = 0.04

# variables that are NOT normally distributed
demo1_val %>% dplyr::select(AGE, LAR, PC20, calculated_Response) %>% 
  gather(DEMO, Value, -calculated_Response) %>% 
  group_by(DEMO, calculated_Response) %>% 
  dplyr::summarise(Median = round(median(Value, na.rm = TRUE), 2), 
    Q1 = quantile(Value, 0.25, na.rm = TRUE), Q3 = quantile(Value, 0.75, na.rm = TRUE))

# variables that are normally distributed
demo1_val %>% dplyr::select(Wt..Kg., HT.cm., BLFEV, PRFEV, EAR, AIS, calculated_Response) %>% 
  gather(DEMO, Value, -calculated_Response) %>% 
  group_by(DEMO, calculated_Response) %>% 
  summarise(Mean = round(mean(Value, na.rm = TRUE), 2), SD = round(sd(Value, na.rm = TRUE), 2))

## post PC20
demo <- readRDS(meta_file_rnapc_dir)
demoPost <- subset(demo, Time == "Post")

demoPostval <- demoPost[paste(as.character(demoPost$NAME), as.character(demoPost$AIC_YMD), sep = "_") %in% paste(as.character(demo1_val$NAME), as.character(demo1_val$AIC_YMD), sep = "_"), ]
table(demoPostval$calculated_Response)

fit <- lm(demoPostval$PC20 ~ demoPostval$calculated_Response)
plot(fit)
shapiro.test(fit$residuals) ## NOT normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = demoPostval$PC20[demoPostval$calculated_Response == "ER"], 
            y = demoPostval$PC20[demoPostval$calculated_Response == "DR"])  # p=0.04

demoPostval %>% dplyr::select(PC20, calculated_Response) %>% 
  gather(DEMO, Value, -calculated_Response) %>% 
  group_by(DEMO, calculated_Response) %>% 
  dplyr::summarise(Median = round(median(Value, na.rm = TRUE), 2), 
    Q1 = quantile(Value, 0.25, na.rm = TRUE), Q3 = quantile(Value, 0.75, na.rm = TRUE))


```


## create cbc dataset and phenotype vectors

```{r}
y.train_cbc  <- factor(demo1_dis$calculated_Response, c("ER", "DR"))
names(y.train_cbc ) <- rownames(demo1_dis)
y.test_cbc  <- factor(demo1_val$calculated_Response, c("ER", "DR"))
names(y.test_cbc ) <- rownames(demo1_val)
table(y.train_cbc ); table(y.test_cbc )

cbc_train <- na.omit(demo1_dis[, c("PC20", "Leukocyte_Counts.x10.9.", "Neu_percent","lym_percent","mono_percent","eos_percent","baso_percent")])
y.train_cbc <- y.train_cbc[rownames(cbc_train)]
cbc_test <- na.omit(demo1_val[, c("PC20", "Leukocyte_Counts.x10.9.", "Neu_percent","lym_percent","mono_percent","eos_percent","baso_percent")])
y.test_cbc <- y.test_cbc[rownames(cbc_test)]

table(y.test_cbc)
## statistical tests for cell-types
## which cell frequencies signficantly changes between er and dr at pre-challenge
## Leukocytes
fit <- lm(cbc_test$Leukocyte_Counts.x10.9. ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals) # normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # P=0.10
## Neutrophils
fit <- lm(cbc_test$Neu_percent ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals) # normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # P=0.04
## Lymphocytes
fit <- lm(cbc_test$lym_percent ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals) # normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # Pt=0.02
## Monocytes
fit <- lm(cbc_test$mono_percent ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals) # normal
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
summary(fit) # Pt=0.62
## Eosinophils
fit <- lm(cbc_test$eos_percent ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals)  # NOT normal!
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = cbc_test$eos_percent[y.test_cbc == "ER"], 
            y = cbc_test$eos_percent[y.test_cbc == "DR"])  # pw=0.70
## Basophils
fit <- lm(cbc_test$baso_percent ~ y.test_cbc)
plot(fit)
shapiro.test(fit$residuals)  # NOT normal!
qqnorm(fit$residuals); qqline(fit$residuals, col = 2)
wilcox.test(x = cbc_test$baso_percent[y.test_cbc == "ER"], 
            y = cbc_test$baso_percent[y.test_cbc == "DR"])  # pw=0.17

# variables that are normally distributed
cbc_test %>% dplyr::select(Leukocyte_Counts.x10.9., Neu_percent, lym_percent, mono_percent) %>% 
  mutate(calculated_Response = as.character(y.test_cbc)) %>% 
  gather(Cells, Value, -calculated_Response) %>% 
  group_by(Cells, calculated_Response) %>% 
  summarise(Mean = round(mean(Value, na.rm = TRUE), 2), SD = round(sd(Value, na.rm = TRUE), 2))

# variables that are NOT normally distributed
cbc_test %>% dplyr::select(eos_percent, baso_percent) %>% 
  mutate(calculated_Response = as.character(y.test_cbc)) %>% 
  gather(Cells, Value, -calculated_Response) %>% 
  group_by(Cells, calculated_Response) %>% 
  dplyr::summarise(Median = round(median(Value, na.rm = TRUE), 2), 
    Q1 = quantile(Value, 0.25, na.rm = TRUE), Q3 = quantile(Value, 0.75, na.rm = TRUE))

```


## Biomarker Panel validation
## compute lambda values for elastic net classifiers

```{r}
set.seed(32)
pos_hk_elasticNet_lambda <- mapply(function(X.train, X.test, y.train, y.test){
  fit <- enet(X = X.train, Y = y.train, alpha = 0, lambda = NULL, 
            family = "binomial", X.test = X.test, Y.test = y.test, 
            filter = "none", topranked = 10, keepVar = NULL, pop.prev = 0.6, cutoff = NULL)
  fit$lambda
}, X.train = pos_hk_demo1_disList, X.test = pos_hk_demo1_valList, y.train = y.train, y.test = y.test)
pos_hk_elasticNet_lambda

```

## Trinity panel
## Enet

```{r}
set.seed(32)
pos_hk_elasticNet <- lapply(1 : length(pos_hk_demo1_disList), function(i){
    fit <- enet(X = pos_hk_demo1_disList[[i]], Y = y.train[[i]], alpha = 0, lambda = pos_hk_elasticNet_lambda[i], 
            family = "binomial", X.test = pos_hk_demo1_valList[[i]], Y.test = y.test[[i]], 
            filter = "none", topranked = 10, keepVar = NULL, pop.prev = 0.6, cutoff = NULL)
  auc <- roc(y.test[[i]], as.numeric(fit$probs), ci = TRUE)
  data <- simple_roc(labels = as.numeric(y.test[[i]])-1, scores = fit$probs)
  data$Panel <- names(pos_hk_demo1_disList)[i]
  data$Classifier <- "Elastic net"
  data$AUC <- paste0("AUROC = ", round(as.numeric(auc$auc), 2))
  data$CI <- paste0("95% CI:", round(as.numeric(auc$ci), 2)[1], "-", round(as.numeric(auc$ci), 2)[3])
  data
})
names(pos_hk_elasticNet) <- names(pos_hk_demo1_disList)
```

### Rf

```{r}
set.seed(3)
pos_hk_rf <- lapply(1 : length(pos_hk_demo1_disList), function(i){
  fit <- rforest(X = pos_hk_demo1_disList[[i]], Y = y.train[[i]],
            family = "binomial", X.test = pos_hk_demo1_valList[[i]], Y.test = y.test[[i]], 
            filter = "none", topranked = 10)
  auc <- roc(y.test[[i]], as.numeric(fit$probs), ci = TRUE)
  data <- simple_roc(labels = as.numeric(y.test[[i]])-1, scores = fit$probs)
  data$Panel <- names(pos_hk_demo1_disList)[i]
  data$Classifier <- "Random forest"
  data$AUC <- paste0("AUROC = ", round(as.numeric(auc$auc), 2))
  data$CI <- paste0("95% CI:", round(as.numeric(auc$ci), 2)[1], "-", round(as.numeric(auc$ci), 2)[3])
  data
})
names(pos_hk_rf) <- names(pos_hk_demo1_disList)



```

## Plot AUROC

```{r}
aucDat <- rbind(do.call(rbind, pos_hk_elasticNet), do.call(rbind, pos_hk_rf))
aucDat$Panel[aucDat$Panel == "ucscGenes"] <- "UCSC genes"
aucDat$Panel[aucDat$Panel == "ucscGeneIso"] <- "UCSC gene-isoforms"
aucDat$Panel[aucDat$Panel == "ensembl"] <- "Ensembl"
aucDat$Panel[aucDat$Panel == "trinity"] <- "Trinity"
aucDat$Panel[aucDat$Panel == "all"] <- "Combined"
aucDat$Panel[aucDat$Panel == "clinical"] <- "Clinical"
aucDat$Panel[aucDat$Panel == "ucscGenes_clinical"] <- "UCSC genes+clinical"
aucDat$Panel[aucDat$Panel == "ucscGeneIso_clinical"] <- "UCSC gene-isoforms+clinical"
aucDat$Panel[aucDat$Panel == "ensembl_clinical"] <- "Ensembl+clinical"
aucDat$Panel[aucDat$Panel == "trinity_clinical"] <- "Trinity+clinical"
aucDat$Panel[aucDat$Panel == "all_clinical"] <- "Combined+clinical"
aucDat$Panel <- factor(aucDat$Panel, 
  levels = c("Clinical", "UCSC genes", "UCSC gene-isoforms", "Ensembl", "Trinity", "Combined",
    "UCSC genes+clinical", "UCSC gene-isoforms+clinical", "Ensembl+clinical", "Trinity+clinical", "Combined+clinical"))
aucDat$PanelType <- unlist(lapply(strsplit(as.character(aucDat$Panel), "\\+"), function(i) i[1]))
aucDat$Cohort <- "nanoString - Validation"
aucDat$ClassifierType <- rep("Molecular", nrow(aucDat))
aucDat$ClassifierType[grep("+clinical", aucDat$Panel)] <- "Molecular+Clinical"
aucDat$ClassifierType[grep("Clinical", aucDat$Panel)] <- "Clinical"
aucDat$PanelType <- factor(aucDat$PanelType, levels = c("Clinical", "UCSC genes", "UCSC gene-isoforms", "Ensembl", "Trinity", "Combined"))

## AUC records
aucLabel <- aucDat %>% dplyr::select(Panel, Classifier, PanelType, ClassifierType, AUC, CI, Cohort) %>% 
            group_by(Panel, Classifier, Cohort) %>% slice(1)
ci <- as.data.frame(do.call(rbind, strsplit(gsub("95% CI:", "", aucLabel$CI), "-")))
aucLabel$min <- as.numeric(as.character(ci$V1))
aucLabel$max <- as.numeric(as.character(ci$V2))
aucLabel$auc <- as.numeric(unlist(lapply(strsplit(aucLabel$AUC, "= "), function(i) i[2])))
aucLabel$ClassifierType <- rep("Molecular", nrow(aucLabel))
aucLabel$ClassifierType[grep("+clinical", aucLabel$Panel)] <- "Molecular+Clinical"
aucLabel$ClassifierType[grep("Clinical", aucLabel$Panel)] <- "Clinical"
aucLabel$x <- 0.75
aucLabel$y <- rep(0.1, nrow(aucLabel))
aucLabel$y[aucLabel$ClassifierType == "Molecular+Clinical"] <- 0.25
aucLabel$y.ci <- rep(0.05, nrow(aucLabel))
aucLabel$y.ci[aucLabel$ClassifierType == "Molecular+Clinical"] <- 0.2
colPalette <- c("dodgerblue", "black")
pdf(paste0(rnaelements_dir, "/results/Figures/Figure4/Figure4b_revised.pdf"), width = 15, height = 6.5)
ggplot(aucDat, aes(x = FPR, y = TPR, color = ClassifierType)) +
  geom_rect(data = subset(aucDat, PanelType == "Trinity"), 
           aes(fill = PanelType),xmin = -Inf,xmax = Inf,
            ymin = -Inf,ymax = Inf, alpha = 0.01, color = "#FC8D62") +
  geom_abline(slope = 1, intercept = 0, linetype = "dashed", col = "black") + 
  geom_point() + geom_line() +
  facet_grid(Classifier ~ PanelType) +
  customTheme(sizeStripFont = 12, xAngle = 0, hjust = 0.5, vjust = 0.5, 
              xSize = 10, ySize = 10, xAxisSize = 10, yAxisSize = 10) +
  #scale_color_manual(values=colPalette, name = "Panel") +
  geom_text(data = aucLabel, aes(x = x, y = y, label = AUC), size = 3) +
  geom_text(data = aucLabel, aes(x = x, y = y.ci, label = CI), size = 3) +
  xlab("False Positive Rate") + ylab("True Positive Rate")
dev.off()


```

## Trinity biomarker fold-changes

```{r}
X <- rbind(pos_hk_demo1_disList$trinity, pos_hk_demo1_valList$trinity)
Y <- c(demo1_dis$calculated_Response, demo1_val$calculated_Response)
colnames(X)[colnames(X) == "unknown1_comp54405_c1_seq1"] <- "TNFRSF10C_intron"
colnames(X)[colnames(X) == "unknown2_comp55647_c0_seq2"] <- "2_LOC101927568_intron"
colnames(X)[colnames(X) == "unknown2a_comp55647_c0_seq2"] <- "2a_LOC105378945_intron"
colnames(X)[colnames(X) == "unknown3_comp56590_c0_seq8"] <- "TNFRSF10C_intron.LOC254896_exon"

ggplotColours <- function(n = 6, h = c(0, 360) + 15){
  if ((diff(h) %% 360) < 1) h[2] <- h[2] - 360/n
  hcl(h = (seq(h[1], h[2], length = n)), c = 100, l = 65)
}
ann_colors <- list(Response = ggplotColours(n=2))
## Figure 4b and c
result <- mixOmics::pca(X, scale = TRUE, center = TRUE)
pdf(rnae_5a_dir, width = 5, height = 4)
plotIndiv(result, ind.names = FALSE, ellipse = TRUE, star = TRUE, ellipse.level = 0.68, 
         group = Y, col.per.group = ann_colors$Response, legend = TRUE, pch = 19, 
  title = "Trinity panel: Individuals")
dev.off()
pdf(rnae_5b_dir, width = 6, height = 4)
plotLoadings(result, title = "Trinity panel: Variable Contribution", size.name = 0.6, size.title = 1)
dev.off()

# Figure 5c
scaledX <- scale(X)
scaledX[which(scaledX < -2)] <- -2
scaledX[which(scaledX > 2)] <- 2
NMF::aheatmap(t(scaledX), annCol = list(Response = Y),
    border_color = "black", annColors = ann_colors, scale = 'row', 
  filename = rnae_5c_dir, width = 12, height = 6)

```

## apply Trinity panel to all subjects of the discovery and validation cohort

```{r}
colMeans(pos_hk_demo1_disList$trinity[y.train$trinity == "DR", ])-colMeans(pos_hk_demo1_disList$trinity[y.train$trinity == "ER", ])

set.seed(121)
## Elastic net
fit <- enet(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity, alpha = pos_hk_elasticNet_lambda["trinity"], lambda = NULL, 
            family = "binomial", X.test = pos_hk_demo1_valList$trinity, Y.test = y.test$trinity, 
            filter = "none", topranked = 10, keepVar = NULL, pop.prev = 0.6, cutoff = NULL)

## using a cut-off of 0.5
pred <- rep(NA, length(fit$probs))
pred[fit$probs >= 0.5] <- "DR"
pred[fit$probs < 0.5] <- "ER"
table(pred=factor(pred, levels(y.test$trinity)), truth=y.test$trinity)
calcPerf(pred=factor(pred, levels(y.test$trinity)), truth=y.test$trinity, prev = 0.6)
# tune cut-off
#fit$perfTest

## random forest
fit <-  rforest(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity,
            family = "binomial", pos_hk_demo1_valList$trinity, Y.test = y.test$trinity, 
            filter = "none", topranked = 10)
## using a cut-off of 0.5
pred <- rep(NA, length(fit$probs))
pred[fit$probs >= 0.5] <- "DR"
pred[fit$probs < 0.5] <- "ER"
table(pred=factor(pred, levels(y.test$trinity)), truth=y.test$trinity)
calcPerf(pred=factor(pred, levels(y.test$trinity)), truth=y.test$trinity, prev = 0.6)
# tune cut-off
#fit$perfTest

```

## apply Trinity panel
### different response

```{r}
## differentiall expression of flippers
all(rownames(demo1_val_repeatDifferentResponse) == rownames(pos_hk_demo1_val_repeatDifferentResponseList$trinity))
group <- factor(demo1_val_repeatDifferentResponse$calculated_Response, c("ER", "DR"))
subject <- as.character(demo1_val_repeatDifferentResponse$NAME)
table(as.character(demo1_val_repeatDifferentResponse$NAME), demo1_val_repeatDifferentResponse$calculated_Response)
unique(subject)

comparisons <- lapply(1 : ncol(pos_hk_demo1_val_repeatDifferentResponseList$trinity), function(i){
  gene <- as.vector(as.matrix(pos_hk_demo1_val_repeatDifferentResponseList$trinity[, i]))
  df <- data.frame(gene, group, subject)
  groupedDat <- groupedData(gene ~ group | subject, data = df)
  fit <- lme(gene ~ group, data = groupedDat, random = ~ 1 | subject, na.action = na.omit)
  coef(summary(fit))[2,]
})

# Elastic net
fit_enet <- enet(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity, alpha = 0, 
                 lambda = pos_hk_elasticNet_lambda["trinity"], 
            family = "binomial", X.test = pos_hk_demo1_val_repeatDifferentResponseList$trinity, Y.test = factor(demo1_val_repeatDifferentResponse$calculated_Response, levels(y.train$trinity)), 
            filter = "none", topranked = 10, keepVar = NULL, pop.prev = 0.6, cutoff = NULL)

Y.test = factor(demo1_val_repeatDifferentResponse$calculated_Response, levels(y.train$trinity))
pred_enet <- rep(NA, length(fit_enet$probs))
pred_enet[fit_enet$probs >= 0.5] <- "DR"
pred_enet[fit_enet$probs < 0.5] <- "ER"
enetPerf_flippers <- calcPerf(pred=factor(pred_enet, levels(Y.test)), truth=Y.test, prev = 0.6)

# random forest
fit_rf <- rforest(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity,
            family = "binomial", X.test = pos_hk_demo1_val_repeatDifferentResponseList$trinity, Y.test = factor(demo1_val_repeatDifferentResponse$calculated_Response, levels(y.train$trinity)), 
            filter = "none", topranked = 10)

Y.test = factor(demo1_val_repeatDifferentResponse$calculated_Response, levels(y.train$trinity))
pred_rf <- rep(NA, length(fit_rf$probs))
pred_rf[fit_rf$probs >= 0.5] <- "DR"
pred_rf[fit_rf$probs < 0.5] <- "ER"
rfPerf_flippers <- calcPerf(pred=factor(pred_rf, levels(Y.test)), truth=Y.test, prev = 0.6)

perf_flippers <- round(rbind(enetPerf_flippers, rfPerf_flippers), 2) %>% as.data.frame %>% 
  mutate(Classifier = c("Elastic net", "Random forest")) %>% 
    gather(Perf, value, -Classifier)
perf_flippers$x <- rep(0.2, nrow(perf_flippers))
perf_flippers$y <- rep(c(-25, -27, -29, -31, -40), each = 2)
perf_flippers$Subject <- "ID"
perf_flippers$Response <- "ER"
perf_flippers$Perf[perf_flippers$Perf == "sens"] <- "Sens"
perf_flippers$Perf[perf_flippers$Perf == "spec"] <- "Spec"
perf_flippers$Perf[perf_flippers$Perf == "npv"] <- "NPV"
perf_flippers$Perf[perf_flippers$Perf == "ppv"] <- "PPV"
perf_flippers$Perf[perf_flippers$Perf == "accuracy"] <- "Accuracy"
perf_flippers$label <- paste(perf_flippers$Perf, "=", paste0(perf_flippers$value))
perf_flippers$Group <- "Different Response"
perf_flippers$Allergen <- "Cat" ## made up

repeatedResponse <- data.frame(prob = c(as.numeric(fit_enet$probs), as.numeric(fit_rf$probs)),
           lar =  c(demo1_val_repeatDifferentResponse$LAR, demo1_val_repeatDifferentResponse$LAR), 
           Subject = c(demo1_val_repeatDifferentResponse$concealed_ID, demo1_val_repeatDifferentResponse$concealed_ID),
           Response = c(demo1_val_repeatDifferentResponse$calculated_Response, demo1_val_repeatDifferentResponse$calculated_Response),
           Classifier = rep(c("Elastic net", "Random forest"), each = nrow(demo1_val_repeatDifferentResponse)),
           Allergen = c(as.character(demo1_val_repeatDifferentResponse$Allergen_cleanLabel), as.character(demo1_val_repeatDifferentResponse$Allergen_cleanLabel)))
repeatedResponse$Group = "Different Response"

```

### repeated DRs

### different response

```{r}
## differentiall expression of flippers
all(rownames(demo1_val_repeatDR) == rownames(pos_hk_demo1_val_repeatDRList$trinity))
group <- factor(demo1_val_repeatDR$calculated_Response, c("ER", "DR"))
subject <- as.character(demo1_val_repeatDR$NAME)
unique(subject)
table(subject); length(subject)

# Elastic net
fit_enet <- enet(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity, alpha = 0, 
                 lambda = pos_hk_elasticNet_lambda["trinity"], 
            family = "multinomial", X.test = pos_hk_demo1_val_repeatDRList$trinity, Y.test = factor(demo1_val_repeatDR$calculated_Response, levels(y.train$trinity)), 
            filter = "none", topranked = 10, keepVar = NULL, pop.prev = 0.6, cutoff = NULL)

Y.test = factor(demo1_val_repeatDR$calculated_Response, levels(y.train$trinity))
pred_enet <- rep(NA, length(fit_enet$probs[,,1][,"DR"]))
pred_enet[fit_enet$probs[,,1][,"DR"] >= 0.5] <- "DR"
pred_enet[fit_enet$probs[,,1][,"DR"] < 0.5] <- "ER"
enetPerf_DRs <- calcPerf(pred=factor(pred_enet, levels(Y.test)), truth=Y.test, prev = 0.6)

# random forest
fit_rf <- rforest(X = pos_hk_demo1_disList$trinity, Y = y.train$trinity,
            family = "multinomial", X.test = pos_hk_demo1_val_repeatDRList$trinity, Y.test = factor(demo1_val_repeatDR$calculated_Response, levels(y.train$trinity)), 
            filter = "none", topranked = 10)

Y.test = factor(demo1_val_repeatDR$calculated_Response, levels(y.train$trinity))
pred_rf <- rep(NA, length(fit_rf$probs))
pred_rf[fit_rf$probs >= 0.5] <- "DR"
pred_rf[fit_rf$probs < 0.5] <- "ER"
rfPerf_DRs <- calcPerf(pred=factor(pred_rf, levels(Y.test)), truth=Y.test, prev = 0.6)

perf_DRs <- round(rbind(enetPerf_DRs, rfPerf_DRs), 2) %>% as.data.frame %>% 
  mutate(Classifier = c("Elastic net", "Random forest")) %>% 
    gather(Perf, value, -Classifier)
perf_DRs$x <- rep(0.2, nrow(perf_flippers))
perf_DRs$y <- rep(c(-25, -27, -29, -31, -40), each = 2)
perf_DRs$Subject <- "ID"
perf_DRs$Response <- "ER"
perf_DRs$Perf[perf_DRs$Perf == "sens"] <- "Sens"
perf_DRs$Perf[perf_DRs$Perf == "spec"] <- "Spec"
perf_DRs$Perf[perf_DRs$Perf == "npv"] <- "NPV"
perf_DRs$Perf[perf_DRs$Perf == "ppv"] <- "PPV"
perf_DRs$Perf[perf_DRs$Perf == "accuracy"] <- "Accuracy"
perf_DRs$label <- paste(perf_DRs$Perf, "=", paste0(perf_DRs$value))
perf_DRs$Group <- "repeated DRs"
perf_DRs$Allergen <- "Cat" ## made up

repeatedDRs <- data.frame(prob = c(as.numeric(fit_enet$probs[,,1][,"DR"]), as.numeric(fit_rf$probs)),
           lar =  c(demo1_val_repeatDR$LAR, demo1_val_repeatDR$LAR), 
           Subject = c(demo1_val_repeatDR$concealed_ID, demo1_val_repeatDR$concealed_ID),
           Response = c(demo1_val_repeatDR$calculated_Response, demo1_val_repeatDR$calculated_Response),
           Classifier = rep(c("Elastic net", "Random forest"), each = nrow(demo1_val_repeatDR)),
           Allergen = c(as.character(demo1_val_repeatDR$Allergen_cleanLabel), as.character(demo1_val_repeatDR$Allergen_cleanLabel)))
repeatedDRs$Group <- "repeated DRs"

```

### Application of the trinity panel to subjects with a different response and repeated DRs

```{r}
pdf(rnae_5d_dir, width = 12, height = 8)
rbind(repeatedResponse, repeatedDRs) %>% 
  ggplot(aes(x = prob, y = lar, group = Subject, color = Allergen)) + 
  geom_line(color = "black") + geom_point(size = 5) + facet_grid(Group~Classifier) +
  customTheme(sizeStripFont = 25, xAngle = 0, hjust = 0.5, vjust = 0.5, 
              xSize = 15, ySize = 15, xAxisSize = 15, yAxisSize = 15) +
  scale_y_continuous(expression('Maximum percent drop in '~ FEV[1]~" (3-7h)")) + 
  xlab("Likelihood of being a dual responder (based on Trinity panel)") +
  geom_hline(yintercept = -15, linetype = "dashed") + geom_vline(xintercept = 0.5, col = "gray") +
  geom_rect(xmin = 0, xmax = 0.5, ymin = -15,ymax = 0, alpha = 0.005, 
            fill = ann_colors$Response[2], color = ann_colors$Response[2]) +
  geom_rect(xmin = 0.5, xmax = 1, ymin = -50,ymax = -15, alpha = 0.005, 
            fill = ann_colors$Response[1], color = ann_colors$Response[1]) + 
  theme(legend.text = element_text(size = 16)) +
  scale_x_continuous(breaks= c(0, 0.25, 0.5, 0.75, 1),
                      labels=c("0", "0.25", "0.5", "0.75", "1")) +
  geom_text(data = subset(rbind(perf_flippers, perf_DRs), Perf == "Accuracy"), 
            aes(x = x, y = y, label = label), col = "black", size = 6) +
  geom_text(x = 0.4, y = -5, label = "ERs", col = ann_colors$Response[2], size = 8) +
  geom_text(x = 0.6, y = -40, label = "DRs", col = ann_colors$Response[1], size = 8)
dev.off()
```
